cometa_ai/src/api/requirements.txt
django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg[binary]==3.2.6
redis==5.0.8
rq==2.2.0
chromadb==0.6.3
PyPDF2==3.0.1
markdown==3.7.0
beautifulsoup4==4.13.3
numpy==2.2.0
chromadb==0.6.3